requirements.txt
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
tokenizers>=0.15.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
einops>=0.7.0
wandb>=0.16.0
tqdm>=4.66.0
